#PAF-50: trimmed down topology operations
